requirements.txt
discord.py==2.3.2
pymongo[srv]==4.6.1